Add unweighted sample (#23)
Both normal and long vectors are supported. The long vector
support including the  corresponding tests is safe-guarded for
32bit machines. Determining random integers within a range is
now part of the RNG wrapper.

For sampling without replacement at low sampling rates a set
based rejection sampling is used, based on a hash- or bitset
depending on the sampling rate. Code duplication between
different vector lengths and different set implementations is
reduced by the use of C++ templates

Should the returned numbers be in the range [1, m] or [0, m-1]?
The former is useful for the one-based vectors in R, the latter
for the zero-based vectors in C++. This offset is configurable
and defaults to the C++ case, since the functions are part
of the C++ API, while the R API uses 'sample.int'.