Allow registering as user-supplied RNG (#67)
* Change RNG initialization

The actual RNG is only seeded in init() instead of (re)created. This allows for calling dqset.seed(NULL) w/o resetting the RNG to the default type.

* Register RNG as user-defined if option dqrng.register_methods is set

The RNG gets initialized by R via user_unif_init() making it unnecessary to call dqset_seed(NULL).

* Unscramble the seed from R

This way set.seed(<num>) and dqset.seed(<num>) give the same result. It is the task of the RNG to initialize its state in a sensible manner.

* Save and restore previous state of RNGkid

* New methods register_methods() and restore_methods() for registering the RNG